requirements.txt
appnope==0.1.2
backcall==0.2.0
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.2
finnomena-api==1.0.0
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.11
jupyter-core==4.7.1
MarkupSafe==1.1.1
numpy==1.20.0
pandas==1.2.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.14.3
prompt-toolkit==3.0.14
ptyprocess==0.7.0
Pygments==2.7.4
pyorbital==1.6.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.2
requests==2.25.1
retrying==1.3.3
scipy==1.6.1
six==1.15.0
tornado==6.1
tqdm==4.57.0
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
gunicorn==19.9.0
streamlit
google-cloud-storage
